+++ src/app/layout/service/layout.service.ts
@@ -1,0 +1,19 @@
+import { Injectable } from '@angular/core';
+import { CommonService } from '../../services/common.services';
+
+@Injectable()
+export class LayoutService {
+
+  constructor(service: CommonService) { }
+
+  public updateAirPollution(): void {
+
+      if (!this.service.airPollutionKeys) {
+
+          this.service.getAirPollution();
+
+      }
+
+  }
+
+}